Remove vueComponent from SheetComponent and code cleanup
